style: fix a lot of clippy issues and refactor
This commit creates a new folder 'algebra', which
will contain everything related to pure maths and
geometry in the world. It also fixed a lot of
small imperfections that clippy warned me about.
