+++ src/acceleration.rs
@@ -1,3 +1,3 @@
-pub mod bounds;
+/// BVHTree acceleration method
 pub mod bvh;
 pub mod queue_systems;
