+++ src/algebra/prelude.rs
@@ -1,0 +1,3 @@
+pub use super::bounds::BoundingBox;
+/// This is the prelude file for the Algebra core
+pub use super::vectors::{Vec2, Vec3, Vertex};
