+++ src/dither.rs
@@ -1,6 +1,6 @@
 #![allow(dead_code)]
 
-use crate::algebra::Vec3;
+use crate::algebra::prelude::*;
 use image::{ImageBuffer, Rgba};
 #[derive(Clone)]
 pub struct ColorPalette(pub Vec<Vec3>);
